Abandon man-colored
I'm pretty sure the reason I did this was not because of a fixation on color,
but because I was getting no formatting at all.  This plugin stopped working
without my noticing.  But NixOS also seems to have gotten nicer manpage
formatting.  Not colorized, but that doesn't matter to me.